Enhance Android app configuration and user experience
- Added lint configuration in `build.gradle.kts` to prevent build failures on specific lint errors and ensure all warnings are displayed.
- Updated `AndroidManifest.xml` to replace the placeholder transcription service with the actual `VoiceInputService`.
- Improved text handling in `HelpActivity` for better compatibility across API levels by adjusting line height settings.
- Enhanced `SettingsActivity` to retrieve the selectable item background from theme attributes for better UI consistency.
- Introduced a new vector drawable for the microphone icon in `VoiceKeyboardView`, replacing the emoji for a more polished appearance.
- Updated `TextProcessor` to include additional patterns for filtering hallucinations and silence in transcriptions.
- Refined test cases in `AppConfigTest` and `TextProcessorTest` to align with updated default values and expected behaviors.

These changes aim to improve code quality, user interface consistency, and overall application reliability.

+++ android/app/src/test/kotlin/com/voiceinput/core/AudioUtilsTest.kt
@@ -33,9 +33,9 @@ class AudioUtilsTest {
         val bytes = AudioUtils.floatToBytes(floats)
 
         val buffer = ByteBuffer.wrap(bytes).order(ByteOrder.LITTLE_ENDIAN)
-        assertEquals(32767, buffer.getShort(0))
-        assertEquals(0, buffer.getShort(2))
-        assertEquals(-32768, buffer.getShort(4))
+        assertEquals(32767.toShort(), buffer.getShort(0))
+        assertEquals(0.toShort(), buffer.getShort(2))
+        assertEquals((-32768).toShort(), buffer.getShort(4))
     }
 
     @Test
